Make Map Work on Web Servers
Google Maps updated their API. The map won't load properly when uploaded to a web server (making it available on a website). This code change fixes this while still maintaining local support. You do have to get a Google Maps API key for uploading it to a web server but they are free with a Google account for the first 25,000 map loads per 24 hours, which isn't bad.